requirements.txt
fastapi[standard]
python-dotenv
aiohttp

--extra-index-url https://download.pytorch.org/whl/cu124
torch
torchvision
torchaudio

# transformers
# accelerate
# huggingface_hub[hf_xet]

qdrant-client
sentence-transformers
pypdf
python-docx

minio

aiogram<4.0

ragas
datasets
psutil
